wind direction function operationsl

+++ Weather.swift
@@ -10,7 +10,6 @@ import UIKit
 
 class Weather {
 
-
     static let currentlyKey = "currently"
     static let temperatureKey = "temperature"
     static let windSpeedKey = "windSpeed"
@@ -26,7 +25,7 @@ class Weather {
     var currently = ""
     var currentTemperature: Float?
     var currentWindSpeed: Float?
-    var currentWindBearing: Int?
+    var currentWindBearing: Float?
     var precipType = "snow"
     var day0Snow: Float?
     var day1Snow: Float?
@@ -47,7 +46,7 @@ class Weather {
             if let windSpeed = currently[Weather.windSpeedKey] as? Float {
                 self.currentWindSpeed = windSpeed
             }
-            if let windBearing = currently[Weather.windBearingKey] as? Int {
+            if let windBearing = currently[Weather.windBearingKey] as? Float {
                 self.currentWindBearing = windBearing
             }
         }
